MemAllocator redesign. vdseFree is written and vdseMalloc is partially rewritten... nothing tested so far..
Note: I've added a bitmap of free/allocated pages to the allocator
to quickly identify if the previous/next group of pages are free
when freeing pages. This enable us to reunite buffers without having
to loop on the linked list of free buffers. 

The cost of the bitmap: 1 bit per page. For 4k pages, this means about
0.003% of the shared memory. In fact for a VDS below ~120 MB, the 
bitmap will fit in the page already allocated to the MemAllocator so the
cost here would be zero... Each additional 125 MB will require adding
an additional page to the allocator.



git-svn-id: https://vdsf.svn.sourceforge.net/svnroot/vdsf/trunk@52 53c26a9f-4021-0410-9f73-c4284c400f72