Fix a bug for win32.
The last argument, a name, in CreateFileMapping is NOT the name of the file but a new name for the resource - giving it the name of the backstore works as long as that name did not include "c:\"... which is not valid for resources. I've set it to null (unnamed) since naming resources like this is not portable and not used in the code.

git-svn-id: https://vdsf.svn.sourceforge.net/svnroot/vdsf/trunk@262 53c26a9f-4021-0410-9f73-c4284c400f72