Reworking on: "Remove the vds temp dir after running the tests.". The last commit did not always work.
This one is better but still not very good. The problem is that the watchdog is not always killed quickly so Sleep() must be added to try to compensate (otherwise, the files are open by vdswd and cannot be removed).

But... how long an interval will depend on the cpu, on how busy the system is, etc. 



git-svn-id: https://vdsf.svn.sourceforge.net/svnroot/vdsf/trunk@576 53c26a9f-4021-0410-9f73-c4284c400f72